safari-600.7-branch/LayoutTests/fast/selectors/lang-conflict-expected.txt

Verify selectors with conflicting :lang().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll(":lang(fr):lang(en)").length is 0
PASS document.querySelectorAll(":lang(en):lang(fr)").length is 0
PASS getComputedStyle(document.getElementById("target")).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
